# packages/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: adres-pl — mikroserwis adresowy
  version: '2.0'
  description: |
    Podpowiedzi adresowe, walidacja i geokodowanie odwrotne na danych
    Państwowego Rejestru Granic (PRG) i TERYT.

    **Uwierzytelnianie: klucz API w nagłówku `X-API-Key`.** Klucz ma postać
    `adr_live_<32 znaki>_<6 znaków sumy>` i jest pokazywany **jeden raz**,
    przy wystawieniu — serwis przechowuje wyłącznie jego skrót.

    Limit zapytań liczony jest po **zweryfikowanym kliencie**, nie po kluczu
    ani po nagłówku: wszystkie klucze jednego klienta dzielą wspólny licznik.
    Ruch bez ważnego klucza podlega osobnemu, niższemu limitowi po adresie
    sieciowym. Limit obowiązuje **na instancję usługi** — przy wielu replikach
    łączna przepustowość jest odpowiednio wyższa.

    Sondy operacyjne (`/health`, `/ready`, `/metrics`, `/status`) pozostają
    bez uwierzytelniania: są potrzebne Kubernetesowi i Prometheuszowi, a odcina
    się je na warstwie sieciowej, nie w aplikacji.

    **Wersja danych** jest niezależna od wersji API. Zwracają ją `/v1/meta`
    oraz pola `wersjaDanych` w odpowiedziach wyszukiwania i walidacji.
  license:
    name: Dane PRG bez licencji (art. 40c ust. 5 Prawa geodezyjnego)
servers:
  - url: http://localhost:3000
    description: Instalacja lokalna

# Obowiazuje wszystkie operacje POZA tymi, ktore maja jawne `security: []`
# (sondy operacyjne). Wpisanie tego globalnie, a nie przy kazdej operacji,
...
# napisac `security: []`, zeby go otworzyc.
security:
  - ApiKey: []
...
paths:
...
  /v1/numbers:
    get:
      tags: [search]
      summary: Numery budynków na ulicy albo w miejscowości bez ulic
      description: |
        Wymagany dokładnie jeden z `ulicId` lub `simc`; brak obu daje 400.
        Zwracamy całą listę i pozwalamy filtrować po stronie klienta — to
        tańsze niż obieg sieciowy przy każdym znaku.
      parameters:
        - { name: ulicId, in: query, schema: { type: integer } }
        - name: simc
          in: query
          description: Dla miejscowości bez ulic. Zwraca wyłącznie punkty bez przypisanej ulicy.
          schema: { type: string, pattern: '^[0-9]{7}$' }
        - { name: prefix, in: query, schema: { type: string, maxLength: 12 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 2000, default: 500 } }
      responses:
        '200':
          description: Numery uszeregowane naturalnie (2 przed 10).
          content:
            application/json:
              schema:
                type: object
                required: [count, results]
                properties:
                  count: { type: integer }
                  results: { type: array, items: { $ref: '#/components/schemas/NumerBudynku' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimitExceeded' }
  /v1/postal-code:
    get:
      tags: [search]
      summary: Kod pocztowy dla adresu
      description: |
        Kod pochodzi z PRG, czyli z ewidencji gminnych — **nie od Poczty
        Polskiej**. Autorytatywny jest Spis PNA (płatny, bez interfejsu
        programistycznego). Rozbieżności są normalne i nie powinny blokować
        zapisu adresu.

        Gdy numeru nie ma w rejestrze, zwracany jest kod dominujący na ulicy
        wraz z polem `uwaga`. Wymagany dokładnie jeden z `ulicId` lub `simc`.
      parameters:
        - { name: nr, in: query, required: true, schema: { type: string, minLength: 1, maxLength: 20 } }
        - { name: ulicId, in: query, schema: { type: integer } }
        - { name: simc, in: query, schema: { type: string, pattern: '^[0-9]{7}$' } }
      responses:
        '200':
          description: Kod pocztowy albo jego przybliżenie.
          content:
            application/json:
              schema:
                type: object
                required: [postalCode, source]
                properties:
                  postalCode: { type: [string, 'null'], example: '00-624' }
                  source:
                    type: string
                    enum: [rejestr_prg, dominujacy_na_ulicy, brak]
                  uwaga:
                    type: string
                    description: Obecne wyłącznie gdy `zrodlo` nie jest `rejestr_prg`.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimitExceeded' }
  /v1/reverse:
    get:
      tags: [search]
      summary: Geokodowanie odwrotne — najbliższe adresy
      description: Współrzędne poza prostokątem obejmującym Polskę dają 400.
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number, minimum: 48.9, maximum: 55.0 } }
        - { name: lon, in: query, required: true, schema: { type: number, minimum: 13.9, maximum: 24.3 } }
        - { name: maxM, in: query, schema: { type: integer, minimum: 1, maximum: 5000, default: 500 } }
      responses:
        '200':
          description: Do pięciu najbliższych punktów, od najbliższego.
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    maxItems: 5
                    items:
                      allOf:
                        - $ref: '#/components/schemas/AdresPelny'
                        - type: object
                          properties:
                            odleglosc_m: { type: number }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimitExceeded' }
  /v1/locality/{simc}:
    get:
      tags: [search]
      summary: Szczegóły miejscowości
      description: |
        Pole `ma_ulice` steruje formularzem: gdy `false`, pole ulicy należy
        **ukryć**, a nie zablokować — numer odnosi się wtedy bezpośrednio
        do miejscowości.
      parameters:
        - { name: simc, in: path, required: true, schema: { type: string, pattern: '^[0-9]{7}$' } }
      responses:
        '200':
          description: Miejscowość wraz z hierarchią administracyjną.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Miejscowosc' }
        '404': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimitExceeded' }
...
components:
  securitySchemes:
    OperatorToken:
      type: http
      scheme: bearer
      description: |
        Token operatora do zarządzania klientami i kluczami. **Odrębny od kluczy
        klienckich**: klucz `adr_live_*` nie otwiera żadnej trasy `/admin` —
        inaczej dowolny klient wystawiłby sobie klucz bez limitów.
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Klucz klienta. **Wyłącznie w nagłówku** — klucz podany w query stringu
        jest traktowany jak jego brak, bo trafiłby do access logu ingressu,
        do nagłówka `Referer` i do historii przeglądarki.
  responses:
...
    Unauthenticated:
      description: |
        Brak klucza albo klucz nieprawidłowy.

        Zły format, zła suma kontrolna i klucz nieznany dają **identyczną**
        odpowiedź — rozróżnienie zamieniłoby ją w wyrocznię dla zgadującego.
      content:
        application/json:
          schema:
            type: object
            required: [error, code]
            properties:
              error: { type: string, example: 'Klucz API nieprawidlowy.' }
              code: { type: string, enum: [BRAK_KLUCZA, NIEPRAWIDLOWY] }
    Forbidden:
      description: |
        Klucz rozpoznany, ale nieczynny. Kod mówi wprost, co się stało —
        na tym etapie klient udowodnił już posiadanie całego sekretu, więc
        nie ma czego chronić, a integrator musi wiedzieć, co zrobić.
      content:
        application/json:
          schema:
            type: object
            required: [error, code]
            properties:
              error: { type: string, example: 'Klucz API wygasl.' }
              code: { type: string, enum: [NIEWAZNY_JESZCZE, WYGASLY, UNIEWAZNIONY, ZAWIESZONY] }
    ValidationError:
      description: Parametry nie przeszły walidacji schematu.
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode: { type: integer, example: 400 }
              code: { type: string, example: FST_ERR_VALIDATION }
              error: { type: string, example: Bad Request }
              message: { type: string, example: 'querystring/simc must match pattern "^[0-9]{7}$"' }
    BadRequest:
      description: Żądanie poprawne składniowo, ale niewykonalne.
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string, example: 'Podaj ulicId albo simc.' }
    RateLimitExceeded:
      description: |
        Przekroczony limit zapytań. Limit liczony jest po adresie sieciowym
        klienta, domyślnie 600 żądań na minutę.
  schemas:
...
    NumerBudynku:
      type: object
      properties:
        id: { type: integer }
        prg_local_id: { type: string, description: Identyfikator rejestrowy. Trwały między zrzutami. }
        nr_budynku: { type: string, example: 12A }
        kod_pocztowy: { type: [string, 'null'] }
        status: { type: [string, 'null'], description: 'Status budynku z PRG. Atrybut znika ze źródła po 1.09.2026 — wartości są zamrożone.' }
        lat: { type: number }
        lon: { type: number }
    Miejscowosc:
      type: object
      properties:
        simc: { type: string }
        name: { type: string }
        kind: { type: integer }
        rodzaj_nazwa: { type: [string, 'null'], example: miasto }
        ma_ulice: { type: boolean }
        liczba_punktow: { type: integer }
        terc_gminy: { type: string }
        gmina: { type: string }
        powiat: { type: [string, 'null'] }
        voivodeship: { type: [string, 'null'] }
        lat: { type: [number, 'null'] }
        lon: { type: [number, 'null'] }
...
    AdresPelny:
      type: object
      additionalProperties: true
      description: Adres z widoku `full_address` — miejscowość, ulica, numer, kod, współrzędne i hierarchia administracyjna.
...